Refactor eqc collision utilities for reuse
