+++ qlbm/components/spacetime/collision/__init__.py
@@ -1,13 +1,13 @@
 """Classes implementing the logic of the collision operator in the Space-Time Data encoding."""
 
-from .eqc_collision import SpaceTimeCollisionOperator
-from .eqc_discretizations import (
-    EquivalenceClass,
+from ....lattice.eqc.eqc import EquivalenceClass
+from .d2q4_old import SpaceTimeD2Q4CollisionOperator
+from ....lattice.eqc.eqc_generator import (
     EquivalenceClassGenerator,
 )
 
 __all__ = [
-    "SpaceTimeCollisionOperator",
+    "SpaceTimeD2Q4CollisionOperator",
     "EquivalenceClass",
     "EquivalenceClassGenerator",
 ]
