+++ qlbm/lattice/eqc/eqc.py
@@ -1,15 +1,11 @@
-"""Equivalence class utility functions. For a discussion of equivalence classes, we refer the reader to Section 4 of :cite:`spacetime2`."""
+from qlbm.lattice.spacetime.properties_base import LatticeDiscretization, LatticeDiscretizationProperties
+from qlbm.tools.exceptions import LatticeException
 
-from itertools import product
-from typing import Dict, List, Set, Tuple, override
 
 import numpy as np
 
-from qlbm.lattice.spacetime.properties_base import (
-    LatticeDiscretization,
-    LatticeDiscretizationProperties,
-)
-from qlbm.tools.exceptions import LatticeException
+
+from typing import List, Set, Tuple, override
 
 
 class EquivalenceClass:
@@ -121,7 +117,7 @@ class EquivalenceClass:
             The mass and momentum of the equivalence class.
         """
         return (self.mass, self.momentum.tolist())
-    
+
     def get_bitstrings(self) -> List[str]:
         """
         Returns the velocity configurations as bitstrings.
@@ -147,58 +143,3 @@ class EquivalenceClass:
     @override
     def __hash__(self):
         return hash((self.discretization, tuple(self.velocity_configurations)))
-
-
-class EquivalenceClassGenerator:
-    """
-    A class that generates equivalence classes for a given lattice discretization.
-
-    .. list-table:: Constructor Attributes
-        :widths: 25 50
-        :header-rows: 1
-
-        * - Attribute
-          - Description
-        * - :attr:`discretization`
-          - The :class:`.LatticeDiscretization` that the equivalence class belongs to.
-
-    """
-
-    discretization: LatticeDiscretization
-
-    def __init__(self, discretization):
-        self.discretization = discretization
-
-    def generate_equivalence_classes(self) -> Set[EquivalenceClass]:
-        """
-        Generates equivalence classes for the given lattice discretization.
-
-        Returns
-        -------
-        Set[EquivalenceClass]
-            All equivalence classes of the discretization.
-        """
-        equivalence_classes: Dict = {}
-        for state in product(
-            [0, 1],
-            repeat=LatticeDiscretizationProperties.get_num_velocities(
-                self.discretization
-            ),
-        ):
-            velocity_vectors = LatticeDiscretizationProperties.get_velocity_vectors(
-                self.discretization
-            )
-            state = np.array(state)  # type: ignore
-            mass = np.sum(state)
-            momentum = np.sum(state[:, None] * velocity_vectors, axis=0)  # type: ignore
-
-            key = (mass, tuple(momentum))
-            if key not in equivalence_classes:
-                equivalence_classes[key] = []
-            equivalence_classes[key].append(state)
-
-        return {
-            EquivalenceClass(self.discretization, set(tuple(cfg.tolist()) for cfg in v))
-            for _, v in equivalence_classes.items()
-            if len(v) > 1
-        }
